- Added ArgParseBuilder + utils file - moved all argparse logic into builder - simplified cli.py file
